\name{bank_income}
\alias{bank_income}
\docType{data}
\title{
Panel data of bank,2001Q1~2010Q1
}
\encoding{latin1}
\description{
A quarterly panel data frame with 1000 observations on the following 7 variables, unbalanced panel data ranges from 2001Q1~2010Q1.
}
\usage{data("bank_income")}

\format{


  \describe{
    \item{\code{ID}}{a numeric vector}
    \item{\code{Qtr}}{a numeric vector}
    \item{\code{preTax_Income}}{a numeric vector}
    \item{\code{shortRatio}}{a numeric vector}
    \item{\code{longRatio}}{a numeric vector}
    \item{\code{Current_ratio}}{a numeric vector}
    \item{\code{LoanDeposit_ratio}}{a numeric vector}
  }
}

\examples{
data(bank_income)
}

