\name{inf19}
\alias{inf19}
\docType{data}
\title{
Monthly inflation time series of 19 countries
}

\description{Monthly inflation time series of 19 countries, 1984.1~2011.3
}
\usage{data(inf19)
}

\format{
  A data frame with 19 countries
  \describe{
    \item{\code{AUSTRIA}}{inflation of Austria}
    \item{\code{BELGIUM}}{inflation of Belgium}
    \item{\code{CANADA}}{inflation of Canada}
    \item{\code{DENMARK}}{inflation of Denmark}
    \item{\code{FINLAND}}{inflation of Finland}
    \item{\code{FRANCE}}{inflation of France}
    \item{\code{GREECE}}{inflation of Greece}
    \item{\code{ICELAND}}{inflation of Iceland}
    \item{\code{ITALY}}{inflation of Italy}
    \item{\code{JAPAN}}{inflation of Japan}
    \item{\code{LUXEMBOURG}}{inflation of Luxembourg}
    \item{\code{NETHERLANDS}}{inflation of Netherlands}
    \item{\code{NORWAY}}{inflation of Norway}
    \item{\code{PORTUGAL}}{inflation of Portugal}
    \item{\code{SPAIN}}{inflation of Spain}
    \item{\code{SWEDEN}}{inflation of Sweden}
    \item{\code{SWITZERLAND}}{inflation of Switzerland}
    \item{\code{UK}}{inflation of UK}
    \item{\code{USA}}{inflation of USA}
    
  }
}

\details{
Monthly CIP, seasonaly differenced of log CPI of 19 countries
}

\examples{
data(inf19)
head(inf19)
}
