\name{inf_Q}
\alias{inf_Q}
\docType{data}
\title{
Quarterly inflation time series of 19 countries
}

\description{Quarterly inflation time series of 19 countries, 1971Q1~2014Q4
}
\usage{data(inf_Q)
}

\format{
  A data frame with 19 countries
  \describe{
    \item{\code{AUSTRLIA}}{inflation of Austrlia}
    \item{\code{AUSTRIA}}{inflation of Austria}
    \item{\code{BELGIUM}}{inflation of Belgium}
    \item{\code{CANADA}}{inflation of Canada}
    \item{\code{DENMARK}}{inflation of Denmark}
    \item{\code{FINLAND}}{inflation of Finland}
    \item{\code{FRANCE}}{inflation of France}
    \item{\code{GREECE}}{inflation of Greece}
    \item{\code{ICELAND}}{inflation of Iceland}
    \item{\code{ITALY}}{inflation of Italy}
    \item{\code{JAPAN}}{inflation of Japan}
    \item{\code{LUXEMBOURG}}{inflation of Luxembourg}
    \item{\code{NETHERLANDS}}{inflation of Netherlands}
    \item{\code{NORWAY}}{inflation of Norway}
    \item{\code{PORTUGAL}}{inflation of Portugal}
    \item{\code{SPAIN}}{inflation of Spain}
    \item{\code{SWEDEN}}{inflation of Sweden}
    \item{\code{SWITZERLAND}}{inflation of Switzerland}
    \item{\code{UK}}{inflation of UK}
    \item{\code{USA}}{inflation of USA}
    
  }
}

\details{
Quarterly CIP, seasonaly differenced of log CPI of 20 countries
}

\examples{
data(inf_Q)
head(inf_Q)
}
