\name{invest}
\alias{invest}
\docType{data}
\title{
investment data of 565 listed companies, 1973-1987
}
\description{investment data of 565 listed companies, 1973-1987, from Hansen's example
}
\usage{data(invest)
}
\format{
  A pooled data frame
  \describe{
    \item{\code{invest[,1]}}{investment/assets}
    \item{\code{invest[,2]}}{Tobin's Q}
    \item{\code{invest[,3]}}{cash-flow/assets}
    \item{\code{invest[,4]}}{debt/assets}
  
  }
}

\details{
This is a pooled data frame, without date (T) and cross-section(N) ID    
columes }


\examples{
data(invest)
head(invest)
}
