\name{sse_calc}
\alias{sse_calc}

\title{a subroutine of model()
}
\description{SSE calculation
}
\usage{
sse_calc(y, x)
}

\arguments{This function is a sub-routine for model(), calculating SSE of each regression
  \item{y}{vector of dependent variable.
}
  \item{x}{matrix of independent variables.
}
}

\references{Hanse B. E. (1999) Threshold effects in non-dynamic panels: Estimation, testing and inference. Journal of Econometrics,93, 345-368.

Original code from Dr. Hansen (http://www.ssc.wisc.edu/~bhansen/).
}

