\name{tr}
\alias{tr}

\title{A sub-routine calculating trace
}
\description{ Estimation of trace.
}
\usage{ tr(y, tt, n)
}

\arguments{This function is a sub-routine for model(), calculating trace  of matrix
  \item{y}{data vector.
}
  \item{tt}{time period length.
}
  
  \item{n}{number of cross-section units.
}
  
}

\references{Hanse B. E. (1999) Threshold effects in non-dynamic panels: Estimation, testing and inference. Journal of Econometrics,93, 345-368.


Original code from Dr. Hansen (http://www.ssc.wisc.edu/~bhansen/).
}
